Reorganization and partial comps decomposition
