+++ frontend/dashboard-mp/src/components/CatalogRegistration/CatalogAuthenticationBinding.js
@@ -1,0 +1,5 @@
+function CatalogAuthenticationBinding(props) {
+
+}
+
+export default CatalogAuthenticationBinding;
