+++ frontend/dashboard-mp/src/components/CatalogRegistration/CatalogForms.js
@@ -1,13 +1,14 @@
 import { useEffect, useState } from "react";
-import { useAuth } from "react-oidc-context";
 import { Accordion, Alert, Button, Card, Col, Container, Form, Row } from "react-bootstrap";
 import { BsFillClipboardFill } from "react-icons/bs";
-import { addServiceProvider } from "../configs/marketplaceConfig";
-import { checkServiceBrokerReadiness } from "../utils/utils";
-import { Front, Github } from 'react-bootstrap-icons';
+import { checkServiceBrokerReadiness } from "../../utils/utils";
+import CatalogDisclaimer from "./CatalogDisclaimer";
+import CatalogInformationsForm from "./CatalogInformationsForm";
+import CatalogAuthenticationBinding from "./CatalogAuthenticationBinding";
+import CatalogRecap from "./CatalogRecap";
+import CatalogRegistrationErrorBar from "./CatalogRegistrationErrorBar";
 
-function ServiceProviderForm(props) {
-    const auth = useAuth();
+function CatalogForms(props) {
     const setIsLoading = props.setIsLoading;
     const [name, setName] = useState("");
     const [description, setDescription] = useState("");
@@ -28,33 +29,7 @@ function ServiceProviderForm(props) {
         console.log("Is loading: " + props.isLoading)
     }, [props.isLoading])
 
-    const handleSubmit = (event) => {
-        event.preventDefault();
-        if (name === "" || description === "" || url === "") {
-            setError("Please fill out all fields!");
-        } else {
-            setError("");
-            setIsLoading(true);
-            checkURL(url).then((error) => {
-                if (error === "") {
-                    console.log("Sending registration request...")
-                    sendRegistrationRequest().then((error) => {
-                        console.log("Registration request sent!")
-                        if (error !== "") {
-                            console.log("Error registering service provider! : " + error)
-                            setError(error);
-                        }
-                        setIsLoading(false);  
-                    });
-                } else {
-                    console.log("Error checking URL! : " + error)
-                    setError(error);
-                    setIsLoading(false);
-                }
-                console.log("Done checking URL!")                
-            });
-        }
-    }
+    
 
     const handleTestSvcProvider = (event) => {
         event.preventDefault();
@@ -76,21 +51,6 @@ function ServiceProviderForm(props) {
         ;
     }
 
-    const sendRegistrationRequest = async () => {
-        // Send HTTP request to backend
-        var response = await addServiceProvider(name, description, url, auth.user?.access_token)
-        if (response.status === 200) {
-            var result = await response.json()
-            console.log("Registration successful! : " + result)
-            setCredentials(result)      
-        } else {
-            var result = await response.json()
-            return "Error registering service provider: " + result.error;
-        }
-
-        return "";
-    }
-
     const reset = () => {
         setName("");
         setDescription("");
@@ -104,19 +64,6 @@ function ServiceProviderForm(props) {
         });
     }
 
-    const checkURL = async (url) => {
-        // Send HTTP request to url + /readyz
-        // If response is 200, return true
-        // Else return false
-        // Check if url is valid
-        try {
-            new URL(url);
-        } catch (_) {
-            return "URL is not valid!";
-        }
-        return "";  
-    }
-
     useEffect(() => {
         if (credentials.authority_url !== "" && credentials.realm !== "" && credentials.client_id !== "" && credentials.client_secret !== "") {
             setDefaultCommand('curl -X POST -H "Content-Type: application/json" -H "X-Broker-API-Version: 2.17" --data \'{"auth_url": "'+credentials.authority_url+'", "realm": "'+credentials.realm+'", "client_id": "'+credentials.client_id+'", "client_secret": "'+credentials.client_secret+'"}\' '+url+'/auth/credentials')
@@ -125,112 +72,56 @@ function ServiceProviderForm(props) {
         }
     }, [credentials])
 
+    switch (step) {
+        case 0:
+            return (
+                <CatalogDisclaimer setStep={setStep}/>
+            )
+        case 1:
+            return (
+                <>
+                    <CatalogRegistrationErrorBar error={error} setError={setError} />
+
+                    <CatalogInformationsForm
+                    setStep={setStep}
+                    setError={setError}
+                    setIsLoading={setIsLoading}
+                    name={name}
+                    setName={setName}
+                    description={description}
+                    setDescription={setDescription}
+                    url={url}
+                    setUrl={setUrl}
+                    setCredentials={setCredentials}
+                    />
+                </>
+                
+            )
+        case 2:
+            return (
+                <>
+                    <CatalogRegistrationErrorBar error={error} setError={setError} />
+
+                    <CatalogAuthenticationBinding setStep={setStep} setError={setError}/>
+                </>
+                
+            )
+        case 3:
+            return (
+                <CatalogRecap setStep={setStep} />
+            )
+        default:
+            return (
+                <>
+                </>
+            )
+    }
+
     if (credentials.authority_url === "" && credentials.realm === "" && credentials.client_id === "" && credentials.client_secret === "") {
         // Show form
         return (
             <>
-            <Container>
-                <Row className="d-flex">
-                    <Col>
-                        <Row>
-                            <Col>
-                                <Container className="p-3 text-center">
-                                    <h1 className="text-center">Catalog Registration</h1>
-                                </Container>
-                            </Col>
-                        </Row>
-                        <Row>
-                            <Col>
-                                <Container fluid className="p-3 ">                        
-                                    <h2>What's a catalog?</h2>
-                                    <p>
-                                        In order to offer your services into the marketplace you can register your catalog in our system.
-                                        The catalog is a REST API server which will expose all your services according to the Open Service Broker API EXTENDED specification.
-                                        To get a seamless communication and functioning with our system, please assure that your catalog meets the requirements.
-                                    </p>
-                                </Container>
-                            </Col>
-                        </Row>
-                        <Row>
-                            <Col>
-                                <Container fluid className="p-3">
-                                    <h3>Requirements</h3>
-                                    <ul>
-                                        <li>Rest API servier compliant to the <a href="https://github.com/andreacv98/service-broker-k8s/blob/keycloak/documentation/spec.md" target="blank">Open Service Broker API extended specification</a> (<a href="https://github.com/andreacv98/service-broker-k8s/blob/keycloak/documentation/openapi.yaml" target="blank">OpenAPI Schema</a>)</li>
-                                        <ul>
-                                            <li>API security supporting <a href="https://openid.net/specs/openid-connect-core-1_0.html" target="blank">OpenID Connect specification</a></li>
-                                            <li><a href="https://liqo.io/" target="blank">Liqo</a> support and integration</li>
-                                        </ul>
-                                        <li>Reachable from Internet network</li>
-                                    </ul>
-                                </Container>
-                            </Col>
-                            <Col>
-                                <Container fluid className="p-3">
-                                    <Card>
-                                        <Card.Header>
-                                            <Card.Title>Catalog example server</Card.Title>
-                                        </Card.Header>
-                                        <Card.Body>
-                                            <Card.Text>
-                                                <p>
-                                                    We can provide you a ready-to-use catalog which meets all the requirements. Just follow the instructions in the <a href="
-                                                    https://github.com/andreacv98/service-broker-k8s/blob/keycloak/README.md" target="_blank">README</a> file of the project.
-                                                    <br/>
-                                                    Some examples can be found in the <a href="https://github.com/andreacv98/service-broker-k8s/tree/keycloak/examples" target="_blank">example</a> folder of the project.
-                                                </p>
-                                            </Card.Text>
-                                        </Card.Body>
-                                        <Card.Footer>
-                                            <Button variant="outline-dark" href="https://github.com/andreacv98/service-broker-k8s" target="_blank" rel="noopener noreferrer">
-                                                <Github size={20} style={{ marginRight: '0.5rem' }} />
-                                                Github repository
-                                            </Button>
-                                        </Card.Footer>
-                                    </Card>                                    
-                                </Container>
-                            </Col>
-                        </Row>
-                    </Col>
-                </Row>
-                <hr />
-                <Row>
-                    <Col>
-                        <h3>1/3 - Catalog informations</h3>
-                    </Col>
-                </Row>
-                <Row>
-                    <Col>
-                        <Form className="p-3">
-                            <Form.Group className="mb-3" controlId="formName">
-                                <Form.Label>Name</Form.Label>
-                                <Form.Control type="text" placeholder="Enter name" value={name} onChange={(e) => setName(e.target.value)} />
-                                <Form.Text>The name of your catalog. It will be recognized by all the users inside the marketplace by this identification.</Form.Text>
-                            </Form.Group>
-                            <Form.Group className="mb-3" controlId="formDescription">
-                                <Form.Label>Description</Form.Label>
-                                <Form.Control as="textarea" placeholder="Enter description" value={description} onChange={(e) => setDescription(e.target.value)} />
-                                <Form.Text>The description of your catalog. Few words to describe the catalog and its content to all users.</Form.Text>
-                            </Form.Group>
-                            <Form.Group className="mb-3" controlId="formUrl">
-                                <Form.Label>URL</Form.Label>
-                                <Form.Control type="text" placeholder="Enter URL" value={url} onChange={(e) => setUrl(e.target.value)} />
-                                <Form.Text>The URL of your catalog. It will be used to reach your catalog and its services. "HTTPS" URL is recommended.</Form.Text>
-                            </Form.Group>
-                            <Button variant="primary" type="submit" onClick={handleSubmit}>Next</Button>
-                        </Form>
-                    </Col>
-                </Row>
-                <Row>
-                    <Col>
-                        <Container className="p-3 text-center">
-                            <Alert variant="danger" show={error !== ""}>
-                                {error}
-                            </Alert>
-                        </Container>
-                    </Col>
-                </Row>
-            </Container>            
+                      
             </>
         )
     } else {
@@ -422,4 +313,4 @@ function ServiceProviderForm(props) {
     
 }
 
-export default ServiceProviderForm;
+export default CatalogForms;
