+++ frontend/dashboard-mp/src/components/CatalogRegistration/CatalogRecap.js
@@ -1,0 +1,5 @@
+function CatalogRecap(props) {
+
+}
+
+export default CatalogRecap;
