+++ frontend/dashboard-mp/src/components/TopBar.js
@@ -18,7 +18,7 @@ function TopBar (props) {
               {
                 auth.isAuthenticated ? (
                   <>
-                    <Nav.Link href="/subscriptions">My purchases</Nav.Link>
+                    <Nav.Link href="/purchases">My purchases</Nav.Link>
                     <Nav.Link href="/deployments">My deployments</Nav.Link>
                   </>
                 ) : (
